backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy[asyncio]==2.0.31
asyncpg==0.29.0
aiosqlite==0.20.0
google-generativeai>=0.5.0
pydantic==2.7.4
python-dotenv==1.0.1
